requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
Django==4.2.9
django-silk==5.1.0
djangorestframework==3.14.0
gprof2dot==2022.7.29
gunicorn==21.2.0
lxml==5.1.0
packaging==23.2
pycodestyle==2.11.1
pytz==2023.3.post1
spyne==2.14.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
